+++ libs/docs/cx/side-navigation/examples/side-navigation-overflow-example.component.ts
@@ -3,7 +3,7 @@ import { CxNestedListModule } from '@fundamental-ngx/cx/nested-list';
 import { CxSideNavigationModule } from '@fundamental-ngx/cx/side-navigation';
 
 @Component({
-    selector: 'fundamental-ngx-side-navigation-overflow-example',
+    selector: 'fdx-side-navigation-overflow-example',
     templateUrl: './side-navigation-overflow-example.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush,
     standalone: true,