+++ libs/datetime-adapter/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0-rc.11](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.10...v0.49.0-rc.11) (2024-02-16)
+
+**Note:** Version bump only for package @fundamental-ngx/datetime-adapter
+
+
+
+
+
+# [0.49.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.9...v0.49.0-rc.10) (2024-02-15)
+
+**Note:** Version bump only for package @fundamental-ngx/datetime-adapter
+
+
+
+
+
 # [0.49.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.49.0-rc.8...v0.49.0-rc.9) (2024-02-15)
 
 **Note:** Version bump only for package @fundamental-ngx/datetime-adapter